-hírdetések megtekintése weblapok felépítése -(előző) weblap endpointjának megalkotása -új osztály (Ad) gyorsabb és könnyebb adatátvitelhez thymelife miatt

+++ src/main/java/iit/unimiskolc/controller/ApiController.java
@@ -43,7 +43,7 @@ public class ApiController {
     }
 
     @RequestMapping("/api/adID/{id}")
-    public List<GameAd> addGame(@PathVariable(value = "id") int id){
+    public List<GameAd> addGame(@PathVariable(value = "id") long id){
         return adService.SearchByID(id);
     }
     @RequestMapping("/api/name/{title}")
